packed_bcd to int

+++ Core/Inc/keypad_module.h
@@ -11,17 +11,17 @@ extern "C" {
 #include "stddef.h"
 #include "stdint.h"
 
-int checkkey(void);
+uint8_t checkkey(void);
 int scankey(void);
 void keypad_ini(void);
 int readkeyboard(int type);
 
-int key_int(void);
+uint8_t key_int(void);
 
 //void write_keypad_lcd(int num);
-int write_keypad_lcd(int fxn,char* num);
+uint8_t write_keypad_lcd(uint8_t fxn, char* num);
 
-int keypad_lcd(int fxn,char* num);
+uint8_t keypad_lcd(uint8_t fxn, char* num);
 
 extern int keynew;
 extern int keyold;
@@ -31,17 +31,17 @@ extern int keyold;
 //  			    KEYPAD 2
 //================================================
 
-int checkkey2(void);
+uint8_t checkkey2(void);
 int scankey2(void);
 void keypad_ini2(void);
 int readkeyboard2(int type);
 
-int key_int2(void);
+uint8_t key_int2(void);
 
 //void write_keypad_lcd(int num);
-int write_keypad_lcd2(int fxn,char* num);
+uint8_t write_keypad_lcd2(uint8_t fxn, char* num);
 
-int keypad_lcd2(int fxn,char* num);
+uint8_t keypad_lcd2(int fxn,char* num);
 
 extern int keynew2;
 extern int keyold2;
